azure-test-env/functions/requirements.txt
azure-functions
azure-functions-durable>=1.2.0
pyodbc
azure-eventhub
azure-storage-blob
azure-ai-formrecognizer>=3.3.0
pandas>=2.0.0
pyarrow>=14.0.0
numpy>=1.24.0
